First pass at removing unsupported/unsupportable
Cookbooks removed in this branch are either deprecated due to
various reasons (projects no longer active or functionality has
been replaced), or because Opscode lacks the capacity to
sufficiently support and maintain them. The cookbooks will still
be available for download from the Community site or from previous
commits in this repository.